+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Delete estimates when assets are archived. [#10](https://github.com/paul121/farm_loocc/issues/10)
 - Add `deleteEstimate($asset)` function to `farm_loocc.estimate` service.
 
 ## 1.0.0 2022-04-02